pyproject.toml
[tool.poetry]
name = "test-swagger-coverage"
version = "2.1.54"
description = "Swagger coverage for API tests"
authors = ["alexanderlozovoy <berpress@gmail.com>"]
packages = [{ include = "swagger_coverage" }]
readme = "README.md"
repository = "https://github.com/berpress/swagger-coverage"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
coverage = "^6.3.2"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
swagger_coverage = "swagger_coverage.scripts.swagger_report:main"
